+++ app/templates/base.jinja
@@ -33,7 +33,7 @@
             </li>
             {% if current_user.is_authenticated %}
             <li class="nav-li hover-check dropdown">
-                <a href={{url_for('bp_main.user_profile_get')}} class="dropdown-link">
+                <a href={{url_for('bp_user.profile_get')}} class="dropdown-link">
                     <img src="{{ url_for('static', filename='nav-bar/user_auth.png') }}" class="user-icon icon-default">
                     <img src="{{ url_for('static', filename='nav-bar/user_hover_auth.png') }}"
                         class="user-icon on-hover">
@@ -41,11 +41,11 @@
                 <div class="dropdown-menu-custom">
                     <div class="dropdown-item-hello">{{current_user.first_name}}, dobrze Cię znowu widzieć!</div>
                     <hr>
-                    <a class="dropdown-item-custom main-navbar-item" href="{{url_for('bp_main.user_profile_get')}}">Moje
+                    <a class="dropdown-item-custom main-navbar-item" href="{{url_for('bp_user.profile_get')}}">Moje
                         dane</a>
                     <a class="dropdown-item-custom main-navbar-item"
-                        href="{{url_for('bp_main.user_settings_get')}}">Ustawienia</a>
-                    <a class="dropdown-item-custom main-navbar-item" href="{{url_for('bp_main.user_offers_get')}}">Moje
+                        href="{{url_for('bp_user.settings_get')}}">Ustawienia</a>
+                    <a class="dropdown-item-custom main-navbar-item" href="{{url_for('bp_user.offers_get')}}">Moje
                         ogłoszenia</a>
                     <a class="dropdown-item-custom main-navbar-item" href="#">Wiadomości</a>
                     <a class="dropdown-item-custom main-navbar-item" href="#">Polubione oferty</a>
@@ -62,7 +62,7 @@
             {% endif %}
             <li class="nav-li button">
                 {% if current_user.is_authenticated %}
-                <a href="{{url_for('bp_main.offer_add')}}">Dodaj ogloszenie</a>
+                <a href="{{url_for('bp_offer.add')}}">Dodaj ogloszenie</a>
                 {% else %}
                 <a href="{{url_for('security.login')}}">Dodaj ogloszenie</a>
                 {% endif %}
