--- app/templates/security/change_password.html
@@ -1,24 +1,0 @@
-{% extends "base.jinja" %}
-{% from "_macros.html" import render_field_with_errors, render_field_errors, render_field %}
-
-{% block style %}
-<link rel="stylesheet" href="{{ url_for('static', filename='styles/login-register.css') }}">
-{% endblock %}
-
-{% block content %}
-<main>
-    <form action="{{ url_for_security('change_password') }}" method="POST" name="change_password_form">
-        <div class="container">
-            <h1>Zmień hasło</h1>
-            <hr>
-            {{ change_password_form.hidden_tag() }}
-            {{ render_field_errors(change_password_form.csrf_token) }} 
-            {{ render_field_with_errors(change_password_form.password) }}
-            {{ render_field_with_errors(change_password_form.new_password) }}
-            {{ render_field_with_errors(change_password_form.new_password_confirm) }}
-            <hr>
-            {{ render_field(change_password_form.submit) }}
-        </div>
-    </form>
-</main>
-{% endblock %}
