+++ app/templates/user/user_profile.jinja
@@ -9,9 +9,9 @@
     <section class="side-navbar-custom">
         <div class="dropdown-item-hello">{{current_user.first_name}}, dobrze Cię znowu widzieć!</div>
         <hr>
-        <a class="dropdown-item-custom" href="{{url_for('bp_main.user_profile_get')}}">Moje dane</a>
-        <a class="dropdown-item-custom" href="{{url_for('bp_main.user_settings_get')}}">Ustawienia</a>
-        <a class="dropdown-item-custom" href="{{url_for('bp_main.user_offers_get')}}">Moje ogłoszenia</a>
+        <a class="dropdown-item-custom" href="{{url_for('bp_user.profile_get')}}">Moje dane</a>
+        <a class="dropdown-item-custom" href="{{url_for('bp_user.settings_get')}}">Ustawienia</a>
+        <a class="dropdown-item-custom" href="{{url_for('bp_user.offers_get')}}">Moje ogłoszenia</a>
         <a class="dropdown-item-custom" href="#">Wiadomości</a>
         <a class="dropdown-item-custom" href="#">Polubione oferty</a>
         <a class="dropdown-item-custom" href="{{ url_for('security.logout') }}">Wyloguj</a>
