+++ app/templates/user/user_settings.jinja
@@ -2,45 +2,61 @@
 {% from "_macros.html" import render_field_with_errors, render_field_errors, render_field %}
 
 {% block style %}
-<link rel="stylesheet" href="{{ url_for('static', filename='styles/user-profile.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='styles/user-profile.css') }}">
 
 {% endblock %}
 
 {% block main_content %}
-<main class>
-    <h1>Ustawienia</h1>
-
-    <section>
-        <form class="form-singleton" action="{{ url_for_security('change_password') }}", method="GET">
-            <input type="submit" value="Zmień hasło" />
-        </form>  
-    </section>
-    <section>
-        <h4>Obecny email: {{ current_user.email }}</h4>
-        <h4>Zmień email</h4>
-        <form action="{{ url_for('bp_main.change_email') }}" method="POST" name="change_email_form">
-            <div class="container">
+    <main class>
+        <h1>Ustawienia</h1>
+        <section>
+            <form action="{{ url_for('security.change_password') }}" method="POST" name="change_password_form">
+                {{ change_password_form.hidden_tag() }}
+                <div class="mb-3">
+                    {{ change_password_form.password.label }}:
+                    {{ change_password_form.password(class='form-control') }}
+                </div>
+                <div class="mb-3">
+                    {{ change_password_form.new_password.label }}:
+                    {{ change_password_form.new_password(class='form-control') }}
+                </div>
+                <div class="mb-3">
+                    {{ change_password_form.new_password_confirm.label }}:
+                    {{ change_password_form.new_password_confirm(class='form-control') }}
+                </div>
+                <div>
+                    {{ change_password_form.submit(class='btn btn-primary') }}
+                </div>
+            </form>
+        </section>
+        <hr>
+        <section>
+            <h4>Obecny email: {{ current_user.email }}</h4>
+            <h4>Zmień email</h4>
+            <form action="{{ url_for('bp_settings.change_email') }}" method="POST" name="change_email_form">
                 {{ change_email_form.hidden_tag() }}
-                {{ render_field_errors(change_email_form.csrf_token) }} 
-                {{ render_field_with_errors(change_email_form.email) }}
-                <hr>
-                {{ render_field(change_email_form.submit) }}
-            </div>
-        </form>
-    </section>
-    <section>
-        <h4>Usuń konto</h4>
-        <form action="{{ url_for('bp_main.delete_account') }}" method="POST" name="delete_account_confirm_form">
-            <div class="container">
-                <p>Potwierdź usunięcie wpisując email</p>
-                <hr>
+                <div class="mb-3">
+                    {{ change_email_form.email.label }}:
+                    {{ change_email_form.email(class='form-control') }}
+                </div>
+                <div>
+                    <button type="submit" class="btn btn-primary">Wyślij link aktywacyjny</button>
+                </div>
+            </form>
+        </section>
+        <hr>
+        <section>
+            <h4>Usuń konto</h4>
+            <form action="{{ url_for('bp_settings.delete_account') }}" method="POST" name="delete_account_confirm_form">
                 {{ delete_account_confirm_form.hidden_tag() }}
-                {{ render_field_errors(delete_account_confirm_form.csrf_token) }} 
-                {{ render_field_with_errors(delete_account_confirm_form.email) }}
-                <hr>
-                {{ render_field(delete_account_confirm_form.submit) }}
-            </div>
-        </form>
-    </section>
-</main>
+                <div class="mb-3">
+                    {{ delete_account_confirm_form.email.label }}:
+                    {{ delete_account_confirm_form.email(class='form-control') }}
+                </div>
+                <div>
+                    <button type="submit" class="btn btn-danger">Usuń konto</button>
+                </div>
+            </form>
+        </section>
+    </main>
 {% endblock %}
